# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Grocery Multi-Vendor API",
        "description": "API documentation for Grocery Multi-Vendor Platform. This API supports Admin, Vendor, Delivery Boy, and Customer roles.",
        "contact": {
            "email": "support@grocery.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List categories",
                "description": "Display a listing of categories",
                "operationId": "e225c2b7eb5daf7fb16e00f4f07ff030",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Filter by parent ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "active_only",
                        "in": "query",
                        "description": "Show only active categories",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create category (Admin only)",
                "description": "Store a newly created category (Admin only)",
                "operationId": "dd76b8d73b7ea8b4951f03d7c0904c92",
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category details",
                "description": "Display the specified category",
                "operationId": "9c4934d1b68a6d4440ec72cfc8ae7074",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category retrieved successfully"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update category (Admin only)",
                "description": "Update the specified category (Admin only)",
                "operationId": "988e8695bc991d7f8e40131db5ba7a76",
                "responses": {
                    "200": {
                        "description": "Category updated successfully"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category (Admin only)",
                "description": "Remove the specified category (Admin only)",
                "operationId": "37196d259228d2274dd9dbef8b00e547",
                "responses": {
                    "200": {
                        "description": "Category deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/categories/{id}/toggle-status": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Toggle category status (Admin only)",
                "description": "Toggle category status (Admin only)",
                "operationId": "7842014721c3e9e23bf7026edccdb1ab",
                "responses": {
                    "200": {
                        "description": "Status updated successfully"
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}